refactor(gui): move the node editor and the shared vocabulary out of flow.ts
The vocabulary — which metrics exist, what a node can be, what a source can be read from, the Modbus
shapes — is the terms every page uses to mean the same thing, so it is not one page's to own. The node
editor is a form, not a diagram: it needs the feeder-loop check from the flow code, and the pages that open
it need three things back.

flow.ts 2749 -> 1954, from 3728 where this started.

The bundler dropped only the first line of an import. A module with a dozen named imports is written across
several lines, so its closing brace stayed in the bundle and nothing parsed — with the error pointing at
the brace, a long way from the cause. It handles both now, and smoke fails if an import statement survives
into the bundle at all.
